x64 fix: moved inline asm intrinsics to separate asm files.

+++ gea/libs/gea_core/src/gea/core/string/string_builder.inl
@@ -52,7 +52,7 @@ inline string_builder_impl<T> &string_builder_impl<T>::append(const T c) {
 
 template <typename T>
 inline string_builder_impl<T> &string_builder_impl<T>::append_format(const T *format, ...) {
-    const size_t remaining = max(m_end - m_pos, 0);
+    const size_t remaining = max(m_end - m_pos, ptrdiff_t(0));
 
     va_list args;
     va_start(args, format);
