+++ gea/libs/gea_math/src/gea/mth_geometry/line2.cpp
@@ -13,7 +13,7 @@ const float line2::length() const
     const pointset_type &points = m_container.m_points;
 
     float length = 0.0f;
-    for (int i = 1, e = points.size(); i < e; ++i) {
+    for (size_t i = 1, e = points.size(); i < e; ++i) {
         length += points[i - 1].to(points[i]).length();
     }
     return length;
